+++ src/p_user.c
@@ -811,7 +811,9 @@ void P_PlayerThink (player_t* player)
 
       // killough 2/8/98, 3/22/98 -- end of weapon selection changes
 
-      if (player->weaponowned[newweapon] && newweapon != player->readyweapon)
+      if (player->weaponowned[newweapon]
+          && (newweapon != player->readyweapon
+              || CASUALPLAY(weapswitch_interruption))) // [Nugget] Weapon-switch interruption
 
         // Do not go to plasma or BFG in shareware,
         //  even if cheated.